v0.15.9.1-dev | minor fixes
